feat: Implement component modularization for flood information page
