+++ src/app/info-banjir/[authorities]/page.jsx
@@ -1,6 +1,5 @@
 "use client";
 
-import Image from "next/image";
 import { useParams } from "next/navigation";
 
 import {
